fix: update metal sheet types and example data

+++ internal/types/metal-sheets.go
@@ -3,7 +3,8 @@ package types
 type MetalSheets []MetalSheet
 
 type MetalSheet struct {
-	DBKey  DBKey            `json:"db_key"`
+	DBKey DBKey `json:"db_key"`
+
 	Format string           `json:"format"`
 	ToolID string           `json:"tool_id"`
 	Data   []MetalSheetData `json:"data"`
@@ -18,15 +19,18 @@ type MetalSheetData struct {
 type MetalSheetDataFilter []uint8
 
 type MetalSheetDataEntry struct {
-	Thickness              uint8                  `json:"thickness"`
-	LowerStampHeight       uint8                  `json:"lower_stamp_height"`
-	LowerStampSheets       uint8                  `json:"lower_stamp_sheets"`
-	UpperStampSheets       uint8                  `json:"upper_stamp_sheets"`
-	ThicknessSettings      uint8                  `json:"thickness_settings"`
+	Thickness float32 `json:"thickness"`
+
+	LowerStampHeight int8 `json:"lower_stamp_height"`
+
+	LowerStampSheets float32 `json:"lower_stamp_sheets"`
+	UpperStampSheets float32 `json:"upper_stamp_sheets"`
+
+	ThicknessSettings      float32                `json:"thickness_settings"`
 	ThicknessSettingsSacmi ThicknessSettingsSacmi `json:"thickness_settings_sacmi"`
 }
 
 type ThicknessSettingsSacmi struct {
-	Max     uint8 `json:"max"`
-	Current uint8 `json:"current"`
+	Max     float32 `json:"max"`
+	Current float32 `json:"current"`
 }
